Allow overwriting of default web preferences
